feat: add metrics for npc count
