+++ apps/server/src/main.ts
@@ -20,14 +20,14 @@ import { serverOptionsSchema } from "./options";
 import { createDBClient } from "./db/client";
 import { ClientRegistry } from "./ClientRegistry";
 import { createRootRouter } from "./modules/router";
-import { collectProcessMetrics } from "./metrics/collectUserMetrics";
+import { collectProcessMetrics } from "./metrics/collectProcessMetrics";
 import { metricsMiddleware } from "./express/metricsMiddleware";
 import { deriveWorldStateForClient } from "./modules/world/deriveWorldStateForClient";
 import { CharacterService } from "./modules/character/service";
 import type { WorldState, WorldSyncServer } from "./modules/world/WorldState";
 import { movementBehavior } from "./traits/movement";
 import { characterRemoveBehavior } from "./modules/character/characterRemoveBehavior";
-import { collectUserMetrics } from "./metrics/collectProcessMetrics";
+import { collectUserMetrics } from "./metrics/collectUserMetrics";
 import { createTickMetricsObserver } from "./metrics/observeTickMetrics";
 import { createExpressLogger } from "./express/createExpressLogger";
 import { createUrlResolver } from "./createUrlResolver";
